Fix failed move of test class
